fix cast of uint to string
It turns out that casting an uint64 to string is allowed, but it does
not produce the number value as a string, but rather it gives a
single-character string with the Unicode character whose codepoint is
determined by the original uint64 value. That's very much not what we
wanted here.

@reimannf FYI