make the scraper more resilient when dealing with failing backends
We have a backend service that has a frustratingly high error rate. (Not
going to call out anyone in particular here, but the commit log should
give it away.) Because the scraper is responsible for initializing some
database entries that the API relies on, the scraper now has a new code
path for creating those entries even when the backend service is down.