allow resources that do not track quota, add "contains" relation to resources
We need this for an upcoming feature where vCPUs and RAM for instances
of different sizes are billed separately and differently. The
"compute/cores" resource will get a "contains" relationship on new
resources like "cores_small/cores_large" and those only report usage,
not quota (and analogously for "compute/ram").