[multi-cluster removal] simplify reports.GetClusters() -> reports.GetCluster()
Most of this is rather straight-forward after the previous removals. The
only thing that I find noteworthy is that several "LEFT OUTER JOIN"
become "JOIN" in the cluster report queries. The outer join was
marginally useful when reporting over multiple clusters to make sure
that empty entries get created for empty clusters, but now that there is
only one cluster, its purpose is entirely gone.