setup reports.GetProjects() with a streaming API
As explained in the code comments, we need to write the JSON response
for ListProjects in a streaming fashion, i.e. one report at a time,
since detailed reports for domains with lots of projects are too large
to fit within the memory budget of our API processes.

This commit only sets up the API for GetProjects() and adjusts the
callers to deal with that API correctly. The adjustments to the internal
workings of GetProjects() will follow in the next commit.