reports.GetProjects(): collect rate data before resource data
This is a prerequisite for switching to a streaming output behavior.

With this change, `?rates=only` now shows `scraped_at` timestamps. It
appears like this was always the intention, but since the rate data
query was using JOIN instead of LEFT OUTER JOIN, `scraped_at` ended up
not being shown in `?rates=only`.