move type OPAConfiguration from cluster config to cluster
It does not come from the config file anymore.

Side-note, I think the loading of OPA policies was broken entirely
within this branch before this commit. The previous version of
SetupOPA() only did something when `cluster.Config.OPA != nil`. Before,
the non-nilness was established by yaml.Unmarshal() finding the config
section, but after the change to environment variables, the section
would always stay nil and therefore never be filled. The tests did not
detect this because they explicitly set the OPA section to non-nil. This
problem should now be resolved.