move plugin-specific config types into the DiscoveryPlugins themselves
On its surface, this is just cleaner separation of concerns. The
pkg/core module should not have to know any types that are only relevant
to the internals of individual plugin implementations.

The bigger picture is that I want to get test coverage for the config
loading and cluster assembling code, by also using it to set up test
scenarios. This means constructing test plugins in the usual way, from
configuration. Without this change, we would have to bleed test-only
types into pkg/core, which I find excessively ugly.

This change is a proof-of-concept for the approach. If this proves to
work, I will also adjust the QuotaPlugin and CapacityPlugin interfaces
in the same way.