capacity-sapcc-ironic: remove getIronicFlavorNames function
We used to need to list flavors in order to get CPU/RAM/disk size
numbers for matching nodes to flavors. Nowadays the matching is done
in Nova/Ironic and presented to us via the resource_class attribute, so
we can fully trust the quota class set on what baremetal flavors exist.