add datamodel.ProjectResourceUpdate
This implementation unifies the create+update logic for
project_resources entries that was previously spread and duplicated
across collector.Scrape, v1Provider.putOrSimulateProject and
datamodel.ValidateProjectServices.

As of this commit, the new implementation is used by scraping and the
project quota PUT endpoint. The consistency check will be moved to it
once I refactor ValidateProjectServices to not duplicate any of the
parts that were moved into ProjectResourceUpdate.

The testcases adjustments are necessary because this change
"accidentally" fixed two bugs:

- Scrape was not updating computed domain quotas when enforcing quota
  constraints on CQD resources.
- PUT was not always enforcing quota constraints.

In the PUT for has_bursting, it's a bit redundant right now that we
use ProjectResourceUpdate to recompute the backend quotas, since
ApplyBackendQuota does it again. I plan to adjust ApplyBackendQuota to
skip the recomputation step and just apply what it finds.