remove Cluster.ServiceTypes field
This field was a sorted duplicate of the list of keys from
Cluster.QuotaPlugins. Because it was a separate field, there was always
the risk that this field could drift out of sync with what's in
QuotaPlugins. Its replacement, the ServiceTypesInAlphabeticalOrder
function, computes the same sorted list at call time.

As a nice demonstration of why the old API was lacking, this change
requires changing test fixtures because some tests set up ServiceTypes
incorrectly (with the correct values, but in the wrong order).