remove all project_resources handling from ValidateProjectServices
Additionally, this removes the unused service list return values from
ValidateProjectServices and also ValidateDomainServices.

The rationale for this change is that all the project_resources handling
is now centralized in ProjectResourceUpdate, and calling this falls
within the duty of the scraper. This change clarifies this separation of
duties (the scraper cares about the project_resources, ScanDomains and
the consistency check care about everything else).