rollup DB migrations
The rolled-up migration exactly matches the current state of our DB
schema. I validated this by running the following before and after the
rollup:

```sh
./testing/with-postgres-db.sh bash -c 'pg_dump -U postgres -h 127.0.0.1 -p 54321 -d limes --schema-only > schema.sql'
```

The output of pg_dump was exactly identical before and after the rollup.

I also used the opportunity to cleanup the whitespace in the SQL code.