capacity-sapcc-ironic: add flavorNamePattern config param
To match the configuration of the quota plugin. This does not cause any
difference in outward behavior because Limes already discards the
capacity values that the capacitor reports for the resources that
`flavorNamePattern` filters away, but the discarding causes a
significant amount of logspam that I would like to get rid of.