allow multiple quota plugins of the same type to be instantiated
By adding an additional `service_type` field to ServiceConfiguration,
same as the `id` field on CapacitorConfiguration.

This change is not relevant for normal plugins, since we only ever have
one of each of those. But in unit tests, we often instantiate multiple
quota plugins of the same type to have multiple services and resources
for different test setups. I want to use the regular config loading
stage during unit tests to get coverage for it, so it will also need to
support this usecase.