make all initialization-phase code return all errors at once
This pattern was already used in some specific places, like parsing of
quota constraints or validation of resource behaviors. Now it extends to
the entire config parsing and initialization code with a uniform
interface.

This also removes all uses of logg.Fatal() from this part of the code,
which is a prerequisite for using it in test setups.