retain gophercloud.ServiceClient instances in plugins
With this change, we only need the ProviderClient/EndpointOpts pair once
during initialization. Afterwards, the plugins maintain their
connections to the respective OpenStack services independently. (Note
that the underlying ProviderClient is still shared between all
ServiceClient instances so constructed, so this does not affect token
rotation and such.)

The main reason why I did not do it like this from the start is that,
very early on, Limes was intended to function with only a partial set of
backend services present (which can happen during buildup of a new
OpenStack cluster). This capability is long since gone, since there are
so many cases now where we query backend services during Init() to
do critical computations like the set of available resources. This
change therefores removed the specific support for missing backend
services from the scrape loops.