parametrize QuotaPlugin interface by service type
The QuotaPlugin interface used to be built on the assumption that
`pluginTypeID == serviceType` for each plugin. This assumption is broken
in the tests, where the same generic plugin gets instantiated multiple
times to test different subvariants in the same test.

The plugin loading code is already aware of the distinction between
`pluginTypeID` and `serviceType`, but while trying to convert the
testcases to load plugins from config, I ran into some places where the
plugin implementations themselves rely on `pluginTypeID == serviceType`.
This commit removes this assumption, by passing in `serviceType` or
`allServiceInfos` from outside where required.