add full AZ awareness to capacity plugins
This is the first half of a refactor of the scrape methods towards full
AZ awareness. Scraping can return data either for the entire region,
or broken down by AZs. This distinction is represented by the enum type
Topological.

For capacity scraping, this does not present a huge change, because
several capacitors already report capacity by AZ. What changes is that:

- subcapacities are now reported on the AZ (instead of regional) level
- the region-wide capacity is computed outside of the plugin by summing
  the individual AZ capacities

The frontend does not change in any way at this point. These changes
will come later, once the user-visible changes have been agreed upon.
This changeset is about preparing the things that we will definitely
need, and making data available to the collector that we will need to
surface in the frontend later on.