add full AZ awareness to quota plugin interface
This is the second part of the change that was started in the previous
commit. This change is much more straight-forward than the one to the
capacity plugins, since quota plugins do not report usage per AZ at all
right now. We are still evaluating which plugins to move to per-AZ usage
reporting and which changes in the backend services will be necessary to
enable this.