correctly distinguish between "unknown usage" and "zero usage" on cluster level
Historically, all cluster resources that had AZ-aware reporting also
reported usage, so it was fine for us to have `usage` be a flat integer.

However, with all resources tracked on AZ level, we need to account for
several non-AZ-aware resources that do not report usage, by making the
usage field a pointer type.