CapacityPluginBackchannel: apply reverse overcommit scaling
Capacity plugins only consider raw capacity, not effective capacity,
since they do not have access to the overcommit factor config. When
we show them resource demand numbers, those are based on the effective
capacity, so we need to apply the raw->effective capacity scaling in
reverse to obtain numbers that are compatible with the capacity that the
plugin sees.