implement new unified theory of Nova capacity calculation
This combines both existing capacitors in a compatible way:
If there are no split flavors at all, the behavior of nova-basic is
recovered. If there are no pooled flavors at all, the behavior of
nova-binpack-flavors is recovered.

The way this is implemented is quite a mess and needs refactoring, but I
want to get it working first because any significant refactor will
require getting rid of the old capacitors beforehand.