rebuild project_commitments with an FK to project_az_resources
This is about the last opportunity to do this sort of easily, before we
get productive data into the project_commitments table. I don't know why
I set up the table with (service_id, resource_id, availability_zone)
rows at all instead of just (az_resource_id). Maybe it's because we only
got the project_az_resources table slightly later? Who knows.

The main purpose of this is to simplify certain OLAP queries, especially
`sum(usage - committed)` over a set of projects (which we need for the
domain and cluster reports). A bit of this is already seen in
AllocationStats and ResourceDemand losing one query each, but the full
extent will become visible in the subsequent commits. Once again, I kept
this initial commit as small as possible for clarity.