break project_resources.{min_quota,max_quota} apart by role
We already have constraints coming from the backend and from the
quota override config, and there is customer demand for a third type of
constraints: Even when project quota is automatically assigned, domain
admins might want to set max_quota to keep costs in check.

This manually-managed constraint clearly needs to be kept separate in
the DB, so all in all, it's the cleanest solution to keep all of them
apart until merging them in ApplyComputedProjectQuota.