confirm available capacity when moving commitments
This fixes a class of bugs where, in the worst case, a malicious user
could create arbitrarily many commitments, even though commitments are
supposed to be bound by the available capacity. The infinite-commitment
glitch would work as follows:

1. Have admin access to two projects (A and B), where project A has some
   non-zero usage of a resource. Project A starts out without
   commitments for that resource.
2. Create an immediately confirmed commitment in project A covering the
   full amount of used resource.
3. Move this commitment to project B.
4. Go back to step 2 and repeat ad nauseam.

The issue here is that the invariant is not just
`sum_over_projects(committed) <= capacity`, but
`sum_over_projects(max(committed, usage)) <= capacity`.

In order to uphold this invariant, it's sometimes necessary to force
commitments to "stick" to the usage that they cover in their existing
project.