capacity-nova: try to respect placements reported by Nova for existing instances
The most significant part here is that, if an existing instance is
placed on a hypervisor that our binpacking simulation ignores because of
"shadowing", we still count this instance into the total capacity.

Shadowing is different from completely ignoring the hypervisor: We want
to entirely ignore HVs with the CUSTOM_DECOMMISSIONING marker trait,
since this is supposed to reduce the capacity on other hypervisors
according to how much usage remains on those HVs.

Shadowing is used with the CUSTOM_HANA_EXCLUSIVE_HOST marker trait. In
regions where we have HANA-exclusive hypervisors, there are still
sometimes legacy instances on general-purpose hypervisors. These legacy
deployments should not block capacity on the exclusive hypervisors
(which currently happens because the legacy instances are placed on the
exclusive hypervisors as part of the binpacking simulation because the
simulation explicitly ignores the general-purpose hypervisors). Hence,
absence of this trait will become a reason for shadowing rather than
ignoring now.