remove domain quota from the DB model
Since this is the only actual data attribute (and not just an identifier
or foreign key) on the domain level, this leads to me removing the
`domain_services` and `domain_resources` tables entirely, which is the
bulk of this PR.

Where domain quota appears in reports, it is replaced by the sum of the
project quotas. This is equivalent to the previous situation because
resources under autogrow already set `domain_quota = sum(project_quota)`
after every ApplyComputedProjectQuota run.