add a separate jobloop for SetQuota
This allows us to run SetQuota more quickly after
ApplyComputedProjectQuota. Before, SetQuota was done as part of
scraping. Since Scrape spends the bulk of the time on *getting* quota
and usage as opposed to *setting* quota, this introduced very
significant delays between quota being distributed on paper and being
actually applied to the backend.