capacity-nova: add shadowing support for pooled resources
The documentation change alludes to what this is about: We are currently
scanning Nova capacity with

```
params:
  required_traits: [ !CUSTOM_DECOMMISIONING, ... ]
  ...
```

This leads to hypervisors that are about to be decommissioned getting
ignored entirely, i.e. they do not count towards capacity at all.
This was fine when we gave out quota manually, and the capacity
operations team was manually keeping track of decom activities.

But now, ConfirmPendingCommitments and ApplyComputedProjectQuota need
a more accurate depiction of available capacity. If there are
hypervisors with the CUSTOM_DECOMMISIONING trait that still have
existing usage on them, it will hurt us to underestimate real capacity
this much. Therefore, it has been decided to count capacity for these
hypervisors **only so far as to cover existing payload on them**.

This is the same logic that we have already implemented for split
flavors under the name "shadowing", so this commit extends the
`shadowing_traits` attribute to work with pooled resources, too.
Then we can change the capacitor config to

```
params:
  shadowing_traits: [ CUSTOM_DECOMMISIONING, ... ]
  ...
```

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch finer-decom-handling
# Changes to be committed:
#	modified:   docs/operators/config.md
#	modified:   internal/plugins/capacity_nova.go
#	modified:   internal/plugins/nova/capacity.go
#