always allow creating commitments to cover existing usage
The previous situation was that, as soon as capacity was overcommitted,
no new commitments could be made at all. Product management decided that
this was undesirable because it locks customers out of discounted prices
for committed resources through no fault of their own.

The function in question previously had test coverage indirectly through
more high-level functions like ConfirmPendingCommitments. For this new
exception, I added a specific testcase because this is easier than
adjusting the high-level tests in the API or the ResourceScrapeJob.