split string identifier types between frontend and backend
Because of the renaming support for ServiceType and ResourceName values
in ResourceBehavior.IdentityInV1API, we have built the appropriate
mappings in *some* places, but not all! To ensure that we do the
conversion every time, in both directions, this introduces separate
string types for the frontend (API) and backend (DB, LIQUID).

On the frontend, we are tied to using the types that appear in
go-api-declarations/limes:

```
limes.ServiceType
limesresources.ResourceName
limesrates.RateName
```

On the backend, we already have `liquid.ResourceName`, which is now not
an alias for `limesresources.ResourceName` anymore. Additional types
are added in `internal/db/`, at the lowest rung of our internal package
hierarchy, leading to the set:

```
db.ServiceType
liquid.ResourceName
db.RateName
```

The latter will become `liquid.RateName` once we add support for rate
scraping in LIQUID. It turned out necessary to do the split for rates
right now because they are grouped into services in the same way that
resources are.

Consequently, QuotaPlugin.Rates gets the same treatment as
QuotaPlugin.Resources and returns a map keyed on `db.RateName` instead
of a slice of objects containing `limesrates.RateName`. This is mostly
for consistency, and also matches what we will likely do in LIQUID once
we add rate scraping support there.

Notes on specific choices:

- internal/core: Name mapping from DB to API (like when building reports
  from DB records) is comparatively cheap because the renaming config is
  structured in this direction, but the reverse lookups are more
  expensive. To avoid incurring too much of a performance penalty, type
  core.NameMapping allows building a full bidirectional mapping that can
  be queried repeatedly in an efficient manner.

- internal/api: Some user-facing error messages do not distinguish
  between "no such service" and "no such resource" anymore because the
  name mapping obscures this division.

- internal/api: I inlined parseAndValidateConversionRequest back into
  its only callsite because I do know how the other conversion endpoint
  will look yet. Please break this back out into a reusable function as
  necessary when adding the second conversion endpoint.

- internal/api: The test data for inconsistency reporting was adjusted
  because it was referring to unconfigured services and resources. This
  used to not be a problem because only the DB was inspected, but now we
  also look at the name mapping.