allow regex replacement in ResourceBehavior.IdentityInV1API
This is somewhere between useful and required for services that have a
large number of resources generated from a pattern, e.g. in Cinder where
there is `{capacity,snapshots,volumes}_$TYPE` for each volume type,
where we may not even know the specific list of volume types ahead of
time.