add core.RateBehavior
This is the same as core.ResourceBehavior, but for rates. I'm adding
this now to support rate renaming in the same way as resource renaming.

There is one clearly ugly part where one specific error message in the
PUT endpoint for rate limits now generates a 500 error in a specific
case that should be a 4xx error. I don't care about fixing this in the
v1 API because nobody is using this at the moment anyway. In the v2 API,
we will transition PUT on rate limits to only work on one rate at the
time in order to simplify the entire RateUpdater thing.