liquid-cinder: fix placement sensing for snapshots on transfered volumes
If a volume was moved to a different project after taking snapshots,
we might only see a snapshot without any volumes associated to it in the
same project. This should not be an error, so we fall back on querying
the individual volume residing in another project with our cloud-admin
privileges.

This was not a problem in the pre-LIQUID implementation because that one
just silently and incorrectly assumed that all snapshots without volume
belonged to the default volume type.