split CommitmentCreationRule/CommitmentConversionRule from ResourceBehavior
Same pattern as in the previous commit.

Based on a review of the existing configuration, it appeared prudent to
split the conversion rules into their own ConfigSet to reduce the amount
of duplication in the config.

This also yields a nice semantic separation, with the remainder of the
commitment configuration now renamed CommitmentCreationRule to clarify
that it only guides the creation (and confirmation) of commitments, not
the maintenance of existing commitments. Consequently, the test for
GC'ing expired commitments does not set up creation rules anymore,
because it ought not need to anyway.