review: do not block renewal on a non-empty transfer status
There is currently no way for a user to reset the transfer status on a
commitment, so if a transfer is started and not completed, this would
block renewing the commitment, which is bad UX.

I understand that this restriction was meant to avoid a race condition
when a commitment is renewed at the same time as being transferred.
But both these operations touch the original commitment, so one of the
associated DB transactions would be forced to fail on commit. That's
good enough of a guarantee for me.