allow domain-specific commitment configuration
Specifically, commitment durations can be set differently per domain
name (and thus, if an empty list of durations is configured for a
domain, that will disable commitments altogether).

CommitmentBehavior is split from ResourceBehavior here, in a move that I
have been considering for a while: ResourceBehavior is a big and bloated
mess, with entries for all kinds of different configurations and
relating to different services all mixed together into a big pile (which
is over 200 LOC in our production config).

I find it more natural to group behaviors under their respective
services, so CommitmentBehavior is the first to move there. Also, it's
the first to use the new regexpext.ConfigSet type, which is aimed at
making sets like ResourceBehavior more type-safe on the code level.

I considered making the full CommitmentBehavior domain-aware for
clarity, i.e. to have the outer type be one of:

    ConfigSet[ResourceName, ConfigSet[DomainName, CommitmentBehavior]]
    ConfigSet[DomainName, ConfigSet[ResourceName, CommitmentBehavior]]

But unfortunately, UntilPercent is categorically global value, and
it also does not really make sense to have different values for
ConversionRule per scope, so I ended up only making the Durations
domain-aware.