adopt type Option as a replacement for pointers
This replaces all fields with pointer types by Option types in LIQUID,
in internal/core/ and internal/db/, where appropriate. I did not touch
the API and report types in order to stay as backwards-compatible as
possible there.

In the v2 API, I would like to see Option types used throughout, and
also for the mess of pointers in the report functions to be trimmed
down.