remove ServiceInfo.ProductName
When the internal service types were "compute", "network" etc., we had
ProductName in order to have "nova", "neutron" etc. on metrics to aid in
alert routing to the respective service channels. Since the DB service
types are now changed to already be "nova", "neutron" etc., we can get
rid of this.