do not fail project deletion on superseded/expired commitments
With the first wave of commitments expiring in prod, we now often see
project deletions failing due to the ON DELETE RESTRICT constraint on
project_commitments. This is generally fine; we do not want to delete a
project that has active commitments. But if the only existing
commitments are in states `superseded` or `expired`, this change will
allow the project deletion to proceed without manual intervention.