simplify CanAcceptCommitmentChanges()
The previous implementation had a somewhat complicated exception case,
where the comment said "what this does is to check that usedCapacity
does not increase". This is a very weird situation: Why use a complex
condition when the condition in the comment is much easier to implement?
This commit does just that.

This came out of a conversation with a customer, who is trying to move
an unused commitment to another project on an AZ resource that is
overbooked (i.e. `usedCapacity > capacity`). Since the commitment is
unused in the donor project, this should be completely fine because
`usedCapacity` does not increase, but the previous implementation
rejected this change because it does not fall into its specific
exception case.