ACPQ: do not block base quota overcommit because of empty AZs in buildup
The addition of the second AZ in Tokyo to our config is causing
problems. There is no compute capacity, but because customers have seen
the new AZ, they tried poking it and creating instances there. These
instances will inevitably fail to spawn because there are no matching
hypervisors to place them on, but the instances are still around in
status `ERROR`. When ACPQ runs, it sees usage in the AZ, but no capacity
to go with it, and so base quota overcommit is blocked.

This is an undesirable behavior, but we don't see a good way to prevent
this situation from arising. Instances in status `ERROR` need to be
reported (because they count towards the quota) and need to be in the
correct AZ (because billing needs to associate them with commitments to
apply the correct price).

We have an existing special case to disregard AZs with zero capacity and
also zero usage for the base quota overcommit (which covers the decom
case), so this gets extended to also include AZs that _never_ had
non-zero capacity (which covers the buildup case).