liquid-neutron: support dynamic quotas
We now support dynamically registered quotas for Neutron, where we only
know a prefix of the quota and not the full name. This is done so
Neutron can expose network flavors as quotas, but might be used for
other objects in the future. To expose these dynamic quotas we're now
checking if the quota name has a prefix from a list of predefined
prefixes (in addition to the lookup in the liquids already existing
hardcoded list).