ensure consistent liquid_version between cluster_services and cluster_resources
These constraints need to be `DEFERRABLE INITIALLY DEFERRED` because
core.Cluster.SaveServiceInfoInDB() needs to update one table first and
then the other afterwards.

As of this commit, the test suite fails because this new constraint
uncovers a bug in the capacity scrape task, where the ClusterService
is being reset to an earlier LiquidVersion after scrapeLiquidCapacity()
runs SaveServiceInfoInDB() in response to a new ServiceInfo version.
The next commit will fix this bug.