add cluster_resources.path and cluster_az_resources.path
For resources, this contains "servicetype/resourcename".
For AZ resources, this contains "servicetype/resourcename/az".

The original intent with this is to make interactive queries easier:
When selecting on project_az_resources, one will only need to join
projects and cluster_az_resources and then select on projects.uuid and
cluster_az_resources.path, without having to join cluster_services and
cluster_resources as well.

While updating the test assertions, I noticed that it also makes
assertions more readable because each line in the SQL diff contains a
full identification of the respective resource or AZ resource.