rename project_commitments.{state => status} and {'active' => 'confirmed'}
This removes the original db.ProjectCommitmentState enum in favor of the
newer liquid.CommitmentStatus enum. In the commitment API, this only
removes a lot of manual type casting because liquid.CommitmentStatus was
already used there throughout.

The only place that still uses the old state enum is the "state" label
of the limes_project_committed_per_az metric. We will clean this up
later with the v2 data metrics.

I also took the opportunity to replace literal mentions of status values
with fmt.Sprintf() in some SQL queries, though admittedly not in all of
them yet.