internal/api: remove one of the start-data fixtures
We have all the tools now to replace this handwritten mess with
something better. I'm starting with the most recent fixture, which is
also one of the smallest, as a field test.

And IMMEDIATELY, this uncovers another bug in the reporting: Once
properly initialized, our domain and project reports show useless empty
entries for AZ "unknown". The cluster level already has a special case
to remove these (just like empty entries for AZ "any"), so we are
applying this fix to the other two levels as well.