move most tests into separate packages
This is the modern best practice because it encourages testing exported
interfaces only (i.e. fully functional subunits instead of isolated
internals). As an exception, two test files in package datamodel stay
with their package because they test a functional core that is not
exposed in the public interface of that package.

In a few places, I had to change private identifiers to public ones.
This ought to be fine because all relevant identifiers are for simple
constants.

This change also enable nice simplifications, coming in the next commit.