remove all hardcoded service/resource/azresource IDs from tests
The upcoming az=total change will shift all of these, and it will be
easier to do this change if the places changed in this commit are
obtained from a dynamic lookup.

I did not touch hardcoded commitment IDs because those will not be
affected by the az=total change. Also, I generally moved towards using
`s.Get{Project,Service,Resource,AZResource}ID()` in order to promote a
consistent style.