internal/core: fix rates records created from configured rates
We only relied on the RateInfo returned by LIQUID when creating `rates`
records, but for rates that are only configured (i.e. limit only and no
usage), we do not have a RateInfo. In this case, we're now putting in
UnitNone and FlatTopology as reasonable fallbacks. If anything else
needs to be supported in the future, we need to extend the config
format.