internal/api: kill fixtures/start-data-commitments.sql
This was less painful than fixtures/start-data.sql in the previous
commit because the project-level data is more uniform. I was surprised,
however, at how significantly the configured ServiceInfo has diverged
from what is in the DB. This should now all be rectified on account of
working from a single source of truth, the ServiceInfo objects
constructed by the test code.