fix data loading order in scrape job
We must defer loading ServiceInfo from the DB until after the scrape.
If the scrape sees a new ServiceInfo version, it will adjust the DB,
which may include substantial changes (e.g. to the set of az_resources
entries, if the topology of a resource changes).

In the next commit, my changes to Test_TopologyScrapes() triggered this
bug because the ServiceInfo changing az-aware -> az-separated was first
observed during Scrape().

In the wild, this bug should not be a big deal, because it appears to
make writeResourceScrapeResult() crash cleanly, thus restarting the
collector and making it observe the ServiceInfo version change during
startup instead.