internal/collector: remove most setup-phase DB test fixtures
In Test_ExpiringCommitmentNotification, I found that many of the
commitments created by the DB fixture were nonsensical (e.g. setting
`confirm_by = expires_at`). While fixing this, I got rid of one of the
expiring short-term commitments because, after fixing it to be
plausible, it ended up being functionally identical to one of the
pending commitments. I also shifted dates around to be plausible, which
slightly altered the AssertEqualf() matches.