collector: detach commitment-related DB test fixtures
It is understandable that TestScanCapacityWithMailNotification() was
originally based on the existing DB fixture for
Test_ScanCapacityWithCommitments(), because setting up a new one used to
be very tedious and not worth it for a small and focused test like this.

However, this caused the problem of making the DB assertions in the
notification test unnecessarily confusing, because commitments
unrelated to the actual test linger in the DB and cause noise from being
confirmed and/or expiring.