collector: remove the last setup-time DB fixture
I'll be the first to admit that I don't particularly like the huge
amount of setup in this test. But it's still better to have this in the
test code itself, with IDs not hardcoded in a DB fixture, and
constructed in a way that has less risk of inconsistencies (e.g.
ExpiresAt is now always calculated to match the specified duration).