add manual test harness for DiscoveryPlugin
We cannot test Keystone API calls with `limesctl liquid`, so we need
something else. This was not a problem for several years because the
Keystone API is usually stable, but today we had a kerfuffle with
a regression related to pagination in Keystone 2025.1, so I started
working on this to diagnose.

The regression was fixed with a config workaround by our Keystone folks,
so no further action is necessary. But since I wrote this test harness,
let's keep it for the next incident.